graph: Move some methods from Schema to ApiSchema
